Custom mutex and semaphore implementations to shrink build size.

+++ utils/memutils.h
@@ -2,10 +2,11 @@
 #define MEMUTILS_H
 
 #include "pattern_scan.h"
+#include "handles.h"
 
 inline uintptr_t GetAbsoluteAddress(uintptr_t addr, intptr_t offset, intptr_t instructionSize)
 {
-	return addr + *(intptr_t*)(addr + offset) + instructionSize;
+	return addr + *(int*)(addr + offset) + instructionSize;
 }
 
 #endif
